Additional tests using @RoutingTest.
