+++ webserver/tests/webserver/src/test/java/io/helidon/webserver/tests/MatchingPatternBase.java
@@ -22,8 +22,6 @@ import io.helidon.webclient.http1.Http1ClientResponse;
 import io.helidon.webserver.http.Handler;
 import io.helidon.webserver.http.HttpRules;
 import io.helidon.webserver.http.HttpService;
-import io.helidon.webserver.testing.junit5.ServerTest;
-import io.helidon.webserver.testing.junit5.SetUpRoute;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -32,29 +30,13 @@ import org.junit.jupiter.params.provider.ValueSource;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
 
-@ServerTest
-class MatchingPatternTest {
-    private static final Handler HANDLER = (req, res) -> res.send(req.matchingPattern().orElse(""));
+abstract class MatchingPatternBase {
+    static final Handler HANDLER = (req, res) -> res.send(req.matchingPattern().orElse(""));
 
     private final Http1Client client;
 
-    MatchingPatternTest(Http1Client client) {
+    MatchingPatternBase(Http1Client client) {
         this.client = client;
-    }
-
-    @SetUpRoute
-    static void routing(HttpRules rules) {
-        rules.get("/", HANDLER)
-                .get("/foo", HANDLER)
-                .get("/foo/bar", HANDLER)
-                .get("/foo/bar/baz", HANDLER)
-                .get("/greet/{name}", HANDLER)
-                .get("/greet/{name1}/{}", HANDLER)
-                .get("/greet/{name1}/and/{name2}", HANDLER)
-                .get("/greet1/*", HANDLER)
-                .get("/greet2/greet/*", HANDLER)
-                .get("/greet3[/greet]", HANDLER)
-                .register("/greet-service", new GreetService());
     }
 
     @ParameterizedTest
